Extract base functionality into an FletcherExtensionArray
Fixes #10
